+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 #Change Log
 All notable changes to this project starting with the 0.6.0 release will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
+## Master
+### Changed
+- Moved Terminus::prompt(), Terminus::promptSecret() to Terminus\Helpers\Input. (#768)
+- Removed duplicative Terminus::menu() in favor of Terminus\Helpers\Input::menu. (#768)
+- Moved Terminus::line() to Terminus\Outputters\Outputter. (#768)
+
 ## [0.10.0] - 2015-12-15
 ### Added
 - New command `workflows show` displays details about a workflow (#687)